+++ src/signal/sources.rs
@@ -6,11 +6,11 @@ use num::Complex;
 #[derive(Debug, Clone)]
 pub struct FromIter<I> {
     iter: I,
-    rate: f64,
+    rate: f32,
 }
 
 impl<I> FromIter<I> {
-    pub fn new(rate: f64, iter: I) -> Self {
+    pub fn new(rate: f32, iter: I) -> Self {
         FromIter {
             iter,
             rate,
@@ -23,50 +23,50 @@ impl<I> Signal for FromIter<I> where I: Iterator {
     fn next(&mut self) -> Option<Self::Sample> {
         self.iter.next()
     }
-    fn rate(&self) -> f64 {
+    fn rate(&self) -> f32 {
         self.rate
     }
 }
 
-pub fn from_iter<I>(rate: f64, iter: I) -> impl Signal<Sample=I::Item>
+pub fn from_iter<I>(rate: f32, iter: I) -> impl Signal<Sample=I::Item>
 where
     I: Iterator,
 {
     FromIter::new(rate, iter)
 }
 
-pub fn from_func<F, A>(rate: f64, mut f: F) -> impl Signal<Sample=A>
+pub fn from_func<F, A>(rate: f32, mut f: F) -> impl Signal<Sample=A>
 where
-    F: FnMut(f64) -> A,
+    F: FnMut(f32) -> A,
 {
     from_iter(rate, Times::new(rate).map(move |t| f(t)))
 }
 
-pub fn constant<A>(rate: f64, value: A) -> impl Signal<Sample=A>
+pub fn constant<A>(rate: f32, value: A) -> impl Signal<Sample=A>
 where
     A: Clone,
 {
     from_iter(rate, std::iter::repeat(value))
 }
 
-pub fn one<A>(rate: f64) -> impl Signal<Sample=A>
+pub fn one<A>(rate: f32) -> impl Signal<Sample=A>
 where
     A: num::One + Clone
 {
     constant(rate, A::one())
 }
 
-pub fn zero<A>(rate: f64) -> impl Signal<Sample=A>
+pub fn zero<A>(rate: f32) -> impl Signal<Sample=A>
 where
     A: num::Zero + Clone
 {
     constant(rate, A::zero())
 }
 
-pub fn freq(rate: f64, freq: f64, phase: f64)
-            -> impl Signal<Sample=Complex<f64>>
+pub fn freq(rate: f32, freq: f32, phase: f32)
+            -> impl Signal<Sample=Complex<f32>>
 {
-    use std::f64::consts::PI;
+    use std::f32::consts::PI;
     from_func(rate,
               move |t| Complex::new(0.0, PI * 2.0 * freq * t + phase).exp())
 }
